New flightmap
Show all graphs in one canvas/image. Prettier graphs with axes.
The map is zoomed/cropped to only show the interesting parts.
Islands are also shown (simple grey polygons).
Graphs can be hovered to show value.
A plane is drawn on top of the route, which can also be hovered.