Add shared settings.ini file
It will be used to set various settings that are of use in both
plugin and web.

In plugin, the preprocess part of the build process will make it
into a header file.

In web, the cli/gensettings.php script will be ran by the Makefile
or on every page load in case $__REPARSE__ is set to true in
inc/conf.php