Generate island map data in separate cli script
It didn't really fit in genislandmaps.
Now it's outputted to a non-specific txt file instead of immediately
formatted for usage in flightmap.js
By this change, the flightmap.php script doesn't need to read and
parse the minimap map files anymore, but can also use the
pregenerated output of this script.