Fix creating trains may use wrong vehicleids
The last implementation of CreateVehicle would create the train
engine and three carriages, but it would take the first available
slot for each of those while it requires 4 sequential ids. The new
implementation will create the train vehicles using a fixed id
instead of letting the vehiclepool decide which id to use.