move static domain to www subfolder /s
It simplifies some stuff.

Using a separate static domain didn't help in this case for cookies
since the website is not served on a www subdomain.

Drawbacks may be that browsers may limit amount of parallel connections
per subdomain and it would be more difficult to let an other http server
serve static resources, but that's worries for when that's actually
needed (which will probably be never).