Use more simpler flight list pagination
The performance was disappointing. Use the previous kind of
pagination, using a 'page' query parameter. This makes the pages
non-deterministic again, but too bad, not a big issue. Keep the
current '< newer | older >' pagination but without the 'newest'
and 'oldest' options, the next idea for a paginator is to use
a giant select à la pouet.net.