Allow filtering on multiple flight states
Instead of only be able to select at most one single flight state,
now multiple can be selected which will behave as an OR condition.

Update the filters design so the filters are hidden and must be
expanded before being able to edit them. This is because the flight
status filter is now a multi-select, so it takes in much more height
than before. The filters form fields would take up too much space on
the page if they were always visible.

Because of this change, it also needs to be clear which filters are
applied because the form fields aren't visible by default. So now
the active filters are shown as text above the form fields, which
doesn't take up as much space.