Combine both flight lists on homepage
Instead of separate sections for 'active flights' and 'last 7 finished
flights', show one section 'recent flights' which is filled with all the
in-progress flights and the 7 most recent not in-progress flights
(so it will not only show successfully landed flights from now on).

It's less bulky like this on the homepage, and 'recent flights' is just
much more simpler to understand. Additionally, the 'active flights'
table was empty most of the time, which didn't look as good.