web: more minor trac edits (mainly styling)
- move some styling to/from trac_inline_style.css
- make breadcrumbs more consistent
- make headers more consistent, prefix with "Tracker: " on subpages
- let form controls decide their own size (except text & textarea)
- apply same column fixes to (some) form tables and changelog tables:
  put the width on 1% with white-space:nowrap
- give errors the "msg error" style

always show the "add comment" section in tracview, but replace the
submit button for a warning when the user is not logged in or doesn't
have perms.
This is because I want to show as much as possible for every user,
and in case something goes wrong (user logs out in different tab),
it means their comment text will be restored into the textarea again
so they won't lose it and can copy it again.
Add an error msg when the form was submitted but rejected.

same for the form on tracnew, but this also meant changing some logic
to allow the page to render (with an extra error message) if the user
is not logged in.

tracview:
- add a "Ticket: " prefix before the summary in the breadcrumbs so
  that the summary can never make the breadcrumbs confusing (for example
  if the ticket is simply named "Change log", it could be confused with
  the actual trac change log page)
- remove the updated timestamp from ticket details table, as this is
  always the timestamp of the last comment anyways
- highlight the "visibility" field if it's not "public"

tracversion:
- fix broken css that changes some margins (blame 7c3cfdd)
- show ticket list as table, show each ticket's reported date