0.3.DL client compat: patch RequestClass
RequestClass got an extra field for custom skin id, so hook SAMP's
packet handler and send data based on whether the client is
0.3.7 or 0.3.DL.

Hooking RequestClass means we can immediately send a random spawn
location in the RequestClass response, meaning sending a random
SpawnInfo in OnPlayerRequestSpawn is not needed anymore, neat!

It also means we don't need to register classes using AddPlayerClass
anymore because we're sending the response ourselves so SAMP doesn't
need to have the spawn info data that it otherwise needs for sending
a response, but we do still need to set the amount of classes that
are available (haven't figured out where that is sent to the
client yet).

Also change SetSpawnInfo because it is different in the same way
(RequestClass just uses SpawnInfo so it's the same)