fix ailerons not visually fixed after /repair
blame 43a51e6 because I immediately changed it to setting vehicle damage
to not-all-zeros while the correct way is to still first set it using
all-zeros and only afterwards setting the vehicle damage to
not-all-zeros. It is a very subtle bug so I suppose nobody has noticed
or assumed it wasn't a bug caused by my code and rather accept it as
a game thing.